+++ packages/fedify/src/federation/mod.ts
@@ -25,7 +25,7 @@ export * from "./mq.ts";
 export type { Message } from "./queue.ts";
 export * from "./retry.ts";
 export * from "./router.ts";
-export { type SenderKeyPair } from "./send.ts";
+export { SendActivityError, type SenderKeyPair } from "./send.ts";
 export {
   handleWebFinger,
   type WebFingerHandlerParameters,